Roles y Permisos Spatie + Middleware en Cultivos

+++ app/Http/Controllers/CropsController.php
@@ -10,13 +10,21 @@ class CropsController extends Controller
 {
     public function index()
     {
-        $crops = Crop::all();
-        return view('crops.index', compact('crops'));
+        if(auth()->user()->can('ver-cultivo'))
+        {
+           $crops = Crop::all();
+           return view('crops.index', compact('crops'));
+        }
+        abort(403);
     }
 
     public function create()
     {
-        return view('crops.create');
+        if(auth()->user()->can('crear-cultivo'))
+        {
+            return view('crops.create');
+        }
+        abort(403);
     }
 
     public function store(Request $request)
@@ -27,14 +35,22 @@ class CropsController extends Controller
 
     public function show($id)
     {
-        $crops = Crop::findOrFail($id);
-        return view('crops.show', compact('crops'));
+        if(auth()->user()->can('ver-cultivo'))
+        {
+            $crops = Crop::findOrFail($id);
+            return view('crops.show', compact('crops'));
+        }
+        abort(403);
     }
 
     public function edit($id)
     {
-        $crops = Crop::findOrFail($id);
-        return view('crops.edit', compact('crops'));
+        if(auth()->user()->can('editar-cultivo'))
+        {
+           $crops = Crop::findOrFail($id);
+           return view('crops.edit', compact('crops'));
+        }
+        abort(403);
     }
 
     public function update(Request $request, $id)
@@ -45,7 +61,11 @@ class CropsController extends Controller
 
     public function destroy($id)
     {
-        Crop::findOrFail($id)->delete();
-        return redirect()->route('crops.index');
+        if(auth()->user()->can('eliminar-cultivo'))
+        {
+          Crop::findOrFail($id)->delete();
+          return redirect()->route('crops.index');
+        }
+        abort(403);
     }
 }
